Rename msvc.DictUnorderedMap to model.DictMap
